Updated the FreeBSD patch to fix build and updated build directions.
Updated the FreeBSD kernel patch so that the FreeBSD Makefiles search for
libsva relative to the source tree and not in a hard coded path relative to
the user's home directory.

Renamed the patch since it is no longer based on a Subversion (SVN) revision
in an internal Subversion repository at University of Illinois.

Rebuilt the patch as a "git diff" patch and updated the README.md file to
reflect the different arguments to the patch command.

Added a clarification to the README.md file on which kernel name is used
in the examples (for users not familiar with the FreeBSD kernel build
procedure).