Simplified the code for identifying the initial thread.
Added a flag to the SVAThread structure to mark that this thread
was created as the first thread for a CPU.

Modified the SVA initialization code and thread allocation code to
correctly configure this flag.

Modified sva_init_stack() to use this flag instead of a flag in
the Interrupt Context to determine if a thread is an initial
thread.

Improved use of blank lines in sva_getCPUState().

There should be no functionality changes.