Tightened the sva_init_stack() state duplication restrictions.
Modified sva_init_stack() so that it does not give child threads
of an initial CPU thread the ability to create their own child
threads.

Modified the code so that it checks the fork bit of the parent
thread and not the child thread.  These should be identical, but
it makes the code more robust to potential future changes and
makes clear what should be done.