Fixed lazy floating point state saving.
Fixed lazy floating point state saving so that the floating point
unit is configured to generate a trap every time we do a context
switch.  This ensures that the kernel or a new thread accessing
the FPU will cause an FP fault and trigger lazy floating point
state saving and restoring.

Removed the old code in the interrupt, trap, and system call
dispatch code that zeros the floating point and vector registers.
With lazy floating point state saving, this code is no longer
needed.