Prevent system call dispatch code from writing to application stack.
Modified the SVA system call dispatch code so that it saves the
initial set of registers within the per-CPU data structure pointed
to by the %gs register instead of saving these registers on the
application stack.

Xiaowan Dong noticed that a page fault when accessing the
application stack causes the SVA trap handler to be in an
inconsistent state (the code is running on a user-space stack but
in kernel mode).  This change fixes that bug.

As this update modified the FreeBSD 9.0 kernel patch, the kernel
must be re-patched in order for this update to work.