Modified method of inserting NOP instructions used for CFI labels.
Modified the CFI instrumentation pass so that it inserts an
X86::NOOP instruction instead of an X86::XCHG32ar instruction
when inserting the latter half of a CFI label.
The LLVM assembler will translate both to the same opcode (0x90),
but the output for the GNU Assembler is different.
This patch ensures that the 0x90 opcode is used regardless of
the method used to assemble the final assembly code.