Moved load_fp() and save_fp() to the state.h header file.
Moved the utility functions load_fp() and save_fp() into the
state.h header file so that there is only one version of each
function that can be inlined into the libsva source code.

Ideally, utility functions should be in header files marked
private (i.e., not included by the kernel), but that is a change
for another time.