Added the -enable-mpx configure script option.
Added an option to the SVA configure script to enable the use of
MPX.

Modified the SVA SFI pass so that, by default, MPX support is
disabled.  An option in make.conf will add a command-line option
to enable MPX support in the compiler when the configure script
is given the option to enable MPX support.