Fixed Issue #6: https://github.com/jtcriswell/SVA/issues/6
This commit modifies the interface to the internal function
unmapSecurePage() so that it takes, as input, a pointer to the
SVAThread for which it should unmap ghost memory.  This fixes
a kernel panic caused by the Virtual Ghost VM attempting to
unmap ghost memory from the currently running SVA thread instead
of the thread ID passed to sva_release_stack().