Enabled the bndPreserve MPX feature to keep bounds-checking enabled.
Added code to init_mpx() to enable the bndPreserve feature.  This
keeps normal call and return instructions from clearing the bounds
registers back to their initial values (thereby disabling bounds
checking and our SFI enforcement).  This appears to decrease
performance but is necessary for correctness.

Modified the range currently configured as valid for kernel code.
This range includes Ghost Memory and SVA VM memory and needs to
be fixed in a later commit.

Modified the testmpx() function to use bounds register 1 (bnd1)
so as not to interefere with the SFI instrumentation (which uses
bnd0).